Fix issues where the SimpleEffectDialog could run callbacks after closing.
With the behaviour of setting a timeout between parameter value changes, this could cause the timeout callback to run sometime after closing the dialog, which can cause confusing behavior. For example, with the Rotate / Zoom dialog this could update the layer's transform to a non-identity value after it was expected to have been reset.

Now, the callback is explicitly invoked (cancelling the timeout) when the dialog is closed, and the Rotate / zoom dialog is destroyed before any further actions.

Fixes: #1905176