Use the standard font chooser button in the text tool.
This should hopefully resolve some existing bugs with large numbers of installed fonts, due to the combobox eagerly rendering all of the font previews. This also allows for easier searching through the fonts, and provides a nice preview of the font and size.

The one downside is opening up a dialog box, but we could avoid that if desired by instead opening a Gtk.Popover with the font chooser widget.

Bug: 1311873, 890589, 1866653