Initial work on fixing some text tool issues with Unicode characters.
There currently are assumptions that each char is a single visible character, which is not the case for surrogate pairs, combining sequences, etc

This change updates the left / right key handlers to use the StringInfo class to advance through text elements rather than chars. The other key hanlders will need similar changes.

Also adjusted the naming of IndexToPosition() and related methods to make it clear that they are dealing with UTF8 byte indices, not (UTF16) char indices or text element indices.

Bug: 1422445