Switch to the standard GTK about dialog.
This reduces a lot of custom code that would need to be maintained when e.g. porting to GTK4

The DLL version info page also isn't as useful now that .NET is bundled with the Mac / Windows installers, but this could be restored in another form if needed.

Moved the application icon (pinta.png / pinta.svg) into Pinta.Resources so that it's handled the same as our other icons, and is included in the Windows / Mac builds. This icon is loaded by the GTK about dialog