Fix a potential crash when pasting into a new image
While creating the new document, the layer list widget resets and triggers HandleLayerSelected() (since the selection resets). This in turn crashed by trying to access the current user layer before the new document had any layers added to it

This can be fixed by not trying to access the document / layers if the selection is being reset, which covers the previous check of HasOpenDocuments. If the selection is being set to something non-empty, then there should be an open document with layers

Fixes: #1959598