Add functions to run dialogs in a blocking manner
In GTK4, dialogs became non-blocking and it is required to connect to the response signal to handle the dialog's action. This is problematic for some existing code related to saving or closing documents, which has complex control flow that expects synchronous behaviour.
These functions help with porting, but in the future we should consider restructuring the code to follow the modern GTK style