Improvements for zooming behaviour
The previous behaviour was to recenter the view around the canvas position that was originally under the scroll wheel. This caused the image to shift around when using the scroll wheel at an off-center position.
Now, the logic is to keep the canvas position under the mouse the same, so that it's easy to Ctrl+scroll to zoom in on a particular area.
When zooming with +/- , the behaviour of zooming around the center of the screen is the same as before.

Also renamed several methods to clarify whether they are working in view space or canvas space

Bug: 1958585