Only add a handler for the brush width's OnValueChanged event once (#1460)
This was being done every time the tool was entered, which is incorrect and will cause the event handler to be run multiple times (after the fixes in gir.core for how closures are implemented).

This also triggered a bug in gir.core from subscribing to the signal multiple times with the exact same delegate.

Bug: #1340