Work around a bug in Mono.Addins with gzip responses (#1547)
Some internal code in Mono.Addins assumes the header's ContentLength property is never null, but this is not the case for gzipped responses from Github's server. This started happening recently so it may have been a change on Github's end.
The workaround is to register our own HttpClient provider which disables gzip compression in our requests.

Fixes: #1542